feat: re-merge whitelist riders PR (#733)
