+++ internal/controller/isbservicerollout/isbservicerollout_controller.go
@@ -1053,6 +1053,12 @@ func (r *ISBServiceRolloutReconciler) GetDesiredRiders(rolloutObject ctlrcommon.
 		unstruc.SetName(fmt.Sprintf("%s-%s", unstruc.GetName(), isbsvcName))
 		desiredRiders = append(desiredRiders, riders.Rider{Definition: unstruc, RequiresProgressive: rider.Progressive})
 	}
+
+	// verify that desiredRiders are all permitted Kinds
+	if !riders.VerifyRiders(desiredRiders) {
+		return desiredRiders, fmt.Errorf("rider definitions contained unpermitted Kind")
+	}
+
 	return desiredRiders, nil
 }
