+++ tests/manifests/special-cases/no-strategy/controller-config.yaml
@@ -19,3 +19,4 @@ data:
         - kind: InterstepBufferService
           schedule: "0,0,10"
       analysisRunTimeout: "1200"
+    permittedRiders: "group=autoscaling.k8s.io,kind=VerticalPodAutoscaler;group=,kind=ConfigMap"
