+++ tests/manifests/special-cases/progressive/controller-config.yaml
@@ -17,3 +17,4 @@ data:
         - kind: InterstepBufferService
           schedule: "0,0,10"
       analysisRunTimeout: "1200"
+    permittedRiders: "group=autoscaling.k8s.io,kind=VerticalPodAutoscaler;group=,kind=ConfigMap"
